ci(GHA): attempt to fix GHA access
For some reason GHA lost access to the ghcr.io
container registry somewhere in the past.

There is a confirmed fix to add GHA write access
to the cloudinfo repository, at the ghcr.io
package settings page, but the
org.opencontainers.image.source label should do
the same when configured correctly.

For now I'm trying to set it to the HTML URL
(HTTPS) instead of the clone URL (SSH).

If this doesn't work we are going to revert it and
settle for the manual write access setting.